Desktop/topic_recsys/requirements.txt
requests
python-dotenv
openai
pendulum>=3.1.0